requirements.txt
Flask==3.0.0
requests==2.31.0
PyPDF2==3.0.1
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.23.0
